requirements.txt
flask
pillow
imagehash
tqdm
pandas
requests
torch
torchvision
transformers
faiss-cpu
numpy
datasets
